feat: 리랜더 최소화 진행
